// apps/admin/src/api/generated/model/feedbackCampaignConversationsDtoOutputCampaignStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type FeedbackCampaignConversationsDtoOutputCampaignStatus =
  (typeof FeedbackCampaignConversationsDtoOutputCampaignStatus)[keyof typeof FeedbackCampaignConversationsDtoOutputCampaignStatus];

export const FeedbackCampaignConversationsDtoOutputCampaignStatus = {
  launched: "launched",
  paused: "paused",
  closed: "closed",
} as const;