// apps/admin/src/api/generated/model/feedbackConversationDetailDtoOutputLifecycleReason.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

/**
 * @nullable
 */
export type FeedbackConversationDetailDtoOutputLifecycleReason =
  | (typeof FeedbackConversationDetailDtoOutputLifecycleReason)[keyof typeof FeedbackConversationDetailDtoOutputLifecycleReason]
  | null;

export const FeedbackConversationDetailDtoOutputLifecycleReason = {
  completed: "completed",
  stopped: "stopped",
  expired: "expired",
  cancelled: "cancelled",
} as const;