// apps/admin/src/api/generated/model/feedbackConversationDetailDtoOutputMessagesItemActor.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type FeedbackConversationDetailDtoOutputMessagesItemActor =
  (typeof FeedbackConversationDetailDtoOutputMessagesItemActor)[keyof typeof FeedbackConversationDetailDtoOutputMessagesItemActor];

export const FeedbackConversationDetailDtoOutputMessagesItemActor = {
  bot: "bot",
  participant: "participant",
  staff: "staff",
  system: "system",
} as const;