// apps/admin/src/api/generated/model/feedbackConversationDetailDtoOutputMessagesItemAttentionCategoriesItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type FeedbackConversationDetailDtoOutputMessagesItemAttentionCategoriesItem =
  (typeof FeedbackConversationDetailDtoOutputMessagesItemAttentionCategoriesItem)[keyof typeof FeedbackConversationDetailDtoOutputMessagesItemAttentionCategoriesItem];

export const FeedbackConversationDetailDtoOutputMessagesItemAttentionCategoriesItem =
  {
    sexual_misconduct: "sexual_misconduct",
    harassment: "harassment",
    violence_or_threat: "violence_or_threat",
    self_harm: "self_harm",
    other_safety: "other_safety",
  } as const;