// apps/admin/src/api/generated/model/feedbackConversationDetailDtoOutputMessagesItemAttentionRecommendedAction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type FeedbackConversationDetailDtoOutputMessagesItemAttentionRecommendedAction =
  (typeof FeedbackConversationDetailDtoOutputMessagesItemAttentionRecommendedAction)[keyof typeof FeedbackConversationDetailDtoOutputMessagesItemAttentionRecommendedAction];

export const FeedbackConversationDetailDtoOutputMessagesItemAttentionRecommendedAction =
  {
    review: "review",
    human_follow_up: "human_follow_up",
    urgent_human_follow_up: "urgent_human_follow_up",
  } as const;