// apps/admin/src/api/generated/model/feedbackConversationResultsDtoOutputAnswersItemQuestionKey.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type FeedbackConversationResultsDtoOutputAnswersItemQuestionKey =
  (typeof FeedbackConversationResultsDtoOutputAnswersItemQuestionKey)[keyof typeof FeedbackConversationResultsDtoOutputAnswersItemQuestionKey];

export const FeedbackConversationResultsDtoOutputAnswersItemQuestionKey = {
  event_score: "event_score",
  liked: "liked",
  meet_again: "meet_again",
  avoid: "avoid",
} as const;