// apps/admin/src/api/generated/model/feedbackConversationResultsDtoOutputNotesItemNoteType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type FeedbackConversationResultsDtoOutputNotesItemNoteType =
  (typeof FeedbackConversationResultsDtoOutputNotesItemNoteType)[keyof typeof FeedbackConversationResultsDtoOutputNotesItemNoteType];

export const FeedbackConversationResultsDtoOutputNotesItemNoteType = {
  activity_interest: "activity_interest",
  general: "general",
} as const;