// apps/admin/src/api/generated/model/feedbackConversationResultsDtoOutputNotesItemOrigin.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type FeedbackConversationResultsDtoOutputNotesItemOrigin =
  (typeof FeedbackConversationResultsDtoOutputNotesItemOrigin)[keyof typeof FeedbackConversationResultsDtoOutputNotesItemOrigin];

export const FeedbackConversationResultsDtoOutputNotesItemOrigin = {
  conversation: "conversation",
  staff: "staff",
} as const;