// apps/admin/src/api/generated/model/feedbackConversationResultsDtoOutputNotesItemStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type FeedbackConversationResultsDtoOutputNotesItemStatus =
  (typeof FeedbackConversationResultsDtoOutputNotesItemStatus)[keyof typeof FeedbackConversationResultsDtoOutputNotesItemStatus];

export const FeedbackConversationResultsDtoOutputNotesItemStatus = {
  new: "new",
  dismissed: "dismissed",
} as const;